+++ react-client/src/components/design-system/index.ts
@@ -1,5 +1,6 @@
 export * from "./button/Button";
 export * from "./button/IconButton";
+export * from "./card/Card";
 export * from "./checkbox/Checkbox";
 export * from "./icon/Icon";
 export * from "./switch/Switch";
