+++ react-client/src/routes/my-pets/MyPetsIndex.tsx
@@ -1,4 +1,4 @@
-import PetCard from "~/components/design-system/card/PetCard";
+import { PetCard } from "~/components/Pet/PetCard";
 import { Button, Title } from "~/components/design-system";
 import { useMyPetsIndexViewModel } from "./useMyPetsIndexViewModel";
 
@@ -24,8 +24,10 @@ export const MyPetsIndex = () => {
   function getHeader() {
     return (
       <div className="mx-xlarge my-large tablet:mx-[0px] tablet:my-[0px]">
-        <div className="tablet:mt-base flex justify-between tablet:px-base">
-          <Title className="text-xl flex items-center tablet:text-2xl">My Pets</Title>
+        <div className="flex justify-between tablet:mt-base tablet:px-base">
+          <Title className="flex items-center text-xl tablet:text-2xl">
+            My Pets
+          </Title>
           <div className="flex">
             <Button
               variant="secondary"
@@ -40,7 +42,7 @@ export const MyPetsIndex = () => {
             </Button>
             <Button
               iconLeft={{ display: "add", size: 16, className: "mr-small" }}
-              className="px-small h-[30px] w-[148px] tablet:w-auto py-small text-[14px] tablet:h-[40px] tablet:py-[12px] tablet:text-[16px]"
+              className="h-[30px] w-[148px] px-small py-small text-[14px] tablet:h-[40px] tablet:w-auto tablet:py-[12px] tablet:text-[16px]"
             >
               Add a new pet
             </Button>
